Add three product flavours in gradle, restart service after reboot, initialize WorkManager via initializer, changes in SendLocationsUseCase

+++ app/src/main/java/ua/com/foxminded/locationtrackera/background/jobs/LocationsUploader.java
@@ -10,17 +10,20 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import javax.inject.Inject;
 
+import io.reactivex.rxjava3.disposables.CompositeDisposable;
 import ua.com.foxminded.locationtrackera.App;
 import ua.com.foxminded.locationtrackera.model.locations.LocationRepository;
 import ua.com.foxminded.locationtrackera.model.usecase.SendLocationsUseCase;
 
-public class LocationsUploader extends Worker implements SendLocationsUseCase.Listener {
+public class LocationsUploader extends Worker {
 
-    private final SendLocationsUseCase sendLocationsUseCase = new SendLocationsUseCase(this);
-    private final AtomicReference<Result> workResult = new AtomicReference<>();
+    private final CompositeDisposable compositeDisposable = new CompositeDisposable();
 
     @Inject
     LocationRepository repository;
+
+    @Inject
+    SendLocationsUseCase sendLocationsUseCase;
 
     public LocationsUploader(@NonNull Context context, @NonNull WorkerParameters workerParams) {
         super(context, workerParams);
@@ -30,24 +33,26 @@ public class LocationsUploader extends Worker implements SendLocationsUseCase.Li
     @NonNull
     @Override
     public Result doWork() {
-        sendLocationsUseCase.execute();
+        final AtomicReference<Result> workResult = new AtomicReference<>();
+        compositeDisposable.add(
+                sendLocationsUseCase.execute()
+                        .subscribe(result -> {
+                            if (result.isSuccessful()) {
+                                repository.deleteLocationsFromDb();
+                                workResult.set(Result.success());
+                            } else {
+                                workResult.set(Result.failure());
+                            }
+                        })
+        );
         return workResult.get();
     }
 
-    @Override
-    public void onLocationsSent() {
-        workResult.set(Result.success());
-        repository.deleteLocationsFromDb();
-    }
-
-    @Override
-    public void onSendingFailed() {
-        workResult.set(Result.failure());
-    }
 
     @Override
     public void onStopped() {
         super.onStopped();
+        compositeDisposable.dispose();
         sendLocationsUseCase.dispose();
     }
 }
